discovery.xml: Disable access rights awareness
This is an alternative method for solving Tomcat and Solr issues
related to users being in large numbers of groups. In the past we
solved this by increasing the Tomcat max HTTP header size and the
Solr max boolean clauses, but we can simply disable access rights
awareness in Discovery. This means that Discovery checks which it-
ems a user has access to before generating the sidebar facets or
community and collection views. As our repository is almost 100%
open access we don't have anything to hide, and anyways if a user
clicks and item they don't have access to then the access rights
will kick in. I imagine this will give Solr a break too.