xmlui: Use haven instead of cookieconsent
For some reason I was never able to figure out the cookieconsent
popup did not get dismissed after clicking agree or disagree. The
cookie was set and the popup did actually work in limiting Google
Analytics to being used only when the user agrees, but the dialog
didn't go away until you navigated to another page.

This new consent manager haven is much nicer to work with and works
by injecting the Google Analytics script and sending the page view
after the user clicks agree. It uses anonymizeIp by default.