xmlui: Implement grunt hack for IE8 JS
As of DSpace 6 Atmire has modified most of the Mirage 2 plumbing to
use npm and yarn for dependency installation. There are two scripts
that are apparently used on old IE (<9) that we don't bundle with
the theme.js and are currently being loaded from node_modules. The
problem is that keeping node_modules in our installation dir means
an extra 200,000 files, which is slow as hell, unneccessary, and
lazy. This introduces a grunt task to copy them from node_modules
to the scripts directory so we can load them from there.