Update hamcrest to v2.2
Due to changes in hamcrest packaging we only need the main artifact
now, but we add hamcrest-core (which is an empty pom) so it doesn't
get pulled in by other deps. Last, the hamcrest docs recommend that
we put hamcrest first so that we don't have dependency convergence
issues from junit.

See: https://hamcrest.org/JavaHamcrest/distributables